requirements.txt
flask
numpy
tflite-runtime
pillow